% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MexBrewer-package.R
\name{ggplot2-scales-continuous}
\alias{ggplot2-scales-continuous}
\alias{scale_color_mex_c}
\alias{scale_colour_mex_c}
\alias{scale_fill_mex_c}
\title{Continuous MexBrewer scales for use with ggplot2}
\usage{
scale_color_mex_c(palette_name, direction = 1, ...)

scale_colour_mex_c(palette_name, direction = 1, ...)

scale_fill_mex_c(palette_name, direction = 1, ...)
}
\arguments{
\item{palette_name}{Name of Palette. Choices are:
\code{Alacena}, \code{Atentado}, \code{Aurora}, \code{Casita1}, \code{Casita2}, \code{Casita3}, \code{Concha}, \code{Frida}, \code{Huida},  \code{Maiz},  \code{Naturaleza},\code{Ofrenda},
\code{Revolucion}, \code{Ronda}, \code{Taurus1}, \code{Taurus2},\code{Tierra}, \code{Vendedora}.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{scale_color_gradientn}}}
}
\value{
A \code{ScaleContinuous} object that can be added to a \code{ggplot} object
}
\description{
Functions \code{\link{scale_color_mex_c}} and \code{\link{scale_fill_mex_c}} for continuous scales enable the use of \code{MexBrewer} colors with \code{ggplot2} continuous scales.
}
\seealso{
Other color scales: 
\code{\link{ggplot2-scales-discrete}}
}
\concept{color scales}
