% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MexBrewer-package.R
\name{met.brewer}
\alias{met.brewer}
\title{Mex Palette Generator}
\usage{
met.brewer(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of Palette. Choices are:
\code{Atentado}, \code{Aurora}, \code{Concha}, \code{Frida}}

\item{n}{Number of desired colors. If number of requested colors is beyond the scope of the palette,
colors are automatically interpolated. If n is not provided, the length of the palette is used.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want to automatically
interpolate between colors.}
}
\value{
A vector of colors.
}
\description{
These are a handful of color palettes from Met.
}
\examples{
mex.brewer("Atentado")
met.brewer("Concha", 6)
met.brewer("Frida", 10, "continuous")
}
\keyword{colors}
