% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MexBrewer-package.R
\docType{package}
\name{MexBrewer-package}
\alias{MexBrewer}
\alias{MexBrewer-package}
\title{MexBrewer: Color Palettes Inspired by Works Mexican Muralists}
\description{
Palettes Inspired by Works at the Mexican Muralists.
}
\author{
\strong{Maintainer}: Antonio Páez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

}
\keyword{internal}
