% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MexBrewer-package.R
\docType{data}
\name{mx_estados}
\alias{mx_estados}
\title{Mexican states.}
\format{
A simple features data frame with 32 rows and 4 variables:
\describe{
\item{ID}{Unique identifier of polygon}
\item{nombre}{Name of the state}
\item{region}{Geographical region of the state; there are five regions in the country}
\item{geometry}{Geometry information of the polygons}
}
}
\usage{
data(mx_estados)
}
\value{
A simple features data frame
}
\description{
A simple features object with the boundaries of states in Mexico (unprojected; CRS is WGS 84).
}
\examples{
 data(mx_estados)
 summary(mx_estados)
}
\keyword{datasets}
