% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MexBrewer-package.R
\name{sequential.palette}
\alias{sequential.palette}
\title{Sequential Palette Check}
\usage{
sequential.palette(palette.name)
}
\arguments{
\item{palette.name}{Name of Palette. Choices are:
\code{Alacena},\code{Atentado}, \code{Aurora}, \code{Concha}, \code{Frida}, \code{Tierra}}
}
\value{
TRUE/FALSE if palette is sequential
}
\description{
Checks whether a palette is Sequential.
}
\examples{
sequential.palette("Aurora")
}
