% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MexBrewer-package.R
\docType{data}
\name{sequential_palettes}
\alias{sequential_palettes}
\title{Names of sequential palettes}
\format{
An object of class \code{character} of length 7.
}
\usage{
sequential_palettes
}
\description{
Use \code{\link{mex.brewer}} to construct palettes.
}
\keyword{datasets}
