% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MexBrewer-package.R
\docType{package}
\name{MexBrewer-package}
\alias{MexBrewer}
\alias{MexBrewer-package}
\title{MexBrewer: Color Palettes Inspired by Works of Mexican Painters and Muralists}
\description{
Color palettes inspired by the works of Mexican painters and muralists. The package includes functions that return vectors of colors and also functions to use color and fill scales in 'ggplot2' visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/MexBrewer}
  \item \url{https://paezha.github.io/MexBrewer/}
  \item Report bugs at \url{https://github.com/paezha/MexBrewer/issues}
}

}
\author{
\strong{Maintainer}: Antonio Páez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

}
\keyword{internal}
