% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MexBrewer-package.R
\docType{data}
\name{MexPalettes}
\alias{MexPalettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 4.
}
\usage{
MexPalettes
}
\description{
Use \code{\link{met.brewer}} to construct palettes.
}
\keyword{datasets}
