% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MexBrewer-package.R
\docType{data}
\name{MexPalettes}
\alias{MexPalettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 18.
}
\usage{
MexPalettes
}
\description{
Use names(MexPalettes) to return all possible palette names. Current choices are:
\code{Alacena}, \code{Atentado}, \code{Aurora}, \code{Casita1}, \code{Casita2}, \code{Casita3}, \code{Concha}, \code{Frida}, \code{Huida},  \code{Maiz}, \code{Naturaleza}, \code{Ofrenda},
\code{Revolucion}, \code{Ronda}, \code{Taurus1}, \code{Taurus2},\code{Tierra}, \code{Vendedora}.
Use \code{\link{mex.brewer}} to construct palettes.
}
\keyword{datasets}
