% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MexBrewer-package.R
\name{sequential.palette}
\alias{sequential.palette}
\title{Sequential Palette Check}
\usage{
sequential.palette(palette_name)
}
\arguments{
\item{palette_name}{Name of Palette. Choices are:
\code{Alacena},\code{Atentado}, \code{Aurora}, \code{Casita1}, \code{Casita2}, \code{Casita3}, \code{Concha}, \code{Frida}, \code{Huida},  \code{Maiz},  \code{Naturaleza},\code{Ofrenda},
\code{Revolucion}, \code{Ronda}, \code{Taurus1}, \code{Taurus2},\code{Tierra}, \code{Vendedora}.}
}
\value{
TRUE/FALSE if palette is sequential
}
\description{
Checks whether a palette is Sequential.
}
\examples{
sequential.palette("Aurora")
}
